Switching over to Heed instead of lmdb-zero.
This allows us to use either lmdb or mdbx, though, I'm currently using
mdbx.

It also allows us to make progress with the iterator that I've been
having issues with.

Added to that, it is much more recently maintained.

Added postcard

Added monotonic increasing LogId for nodes added in the same
millisecond.

Added other few things, and basically got this working like before we
moved to heed.

Next I need to actually add edges back, and fix the examples to use the iterator