Adding transaction timeout
This should make it support a single process creating write
transactions. We now need to support multiple processes creating
transactions simultaneously, and properly handling the mutexes and the timeouts...

Lots of ugly unsafe code, but I'm fairly confident I've upheld the
invariants, for the moment. Of course, this requires better testing to
properly verify.